cortex/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
python-dateutil>=2.8.0 